High_concurrency_flash_sale_system/requirements.txt
fastapi~=0.135.2
SQLAlchemy~=2.0.48
pydantic-settings~=2.13.1
python-jose~=3.5.0
passlib~=1.7.4
pydantic~=2.12.5